mockup_rag/pyproject.toml
[project]
name="mockup-rag"
version="0.0.1"
dependencies=[
    "opensearch_py==2.6.0",
    "opensearch_py_ml==1.1.0",
    "pandas==2.2.2",
    "cohere==5.6.2",
    "requests==2.32.3",
    "termcolor==2.4.0",
    "python-dotenv==1.0.1",
    "tqdm==4.66.4",
    "fastapi==0.111.1",
]
[tool.setuptools]
packages=["mockup_rag"]